validation for delete task

+++ src/app/todo-task/todo-task.component.ts
@@ -1,6 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {TasksService} from '../services/tasks.service';
 import {Task} from '../models/task';
+import {DialogService} from "../services/dialog.service";
 
 @Component({
   selector: 'app-todo-task',
@@ -11,7 +12,7 @@ export class TodoTaskComponent implements OnInit {
 
   tasksList: Array<Task> = [];
 
-  constructor(private tasksService: TasksService) {
+  constructor(private tasksService: TasksService, private dialogService: DialogService) {
     this.tasksService.getTasksListObs().subscribe((tasks: Array<Task>) => {
       this.tasksList = tasks.filter(t => t.isDone === false);
     });
@@ -21,7 +22,14 @@ export class TodoTaskComponent implements OnInit {
   }
 
   remove(task: Task) {
-    this.tasksService.remove(task);
+    const msg = 'Are you sure you want delete task?  ';
+    this.dialogService.openConfirmDialog(msg)
+      .afterClosed().subscribe(resp => {
+       if (resp){
+         this.tasksService.remove(task);
+       }
+    });
+
   }
 
   done(task: Task) {
